Added Serialization Support for Binary Data (i.e. images)
This is a work in progress.  I added a new ORBinaryData type, which is a wrapper for NSData.  If a model uses this type, its binary data (images) will either be serialized inline using Base64 encoding or captured by the serializer to an attachments list, so that the caller can include it as part of a multipart/related upload.  Right now, this new feature is only supported in the XML serializer, and it is uni-directional (i.e. images can be serialized, but not deserialized.)